Admin init, routes, EmptyMessage, etc.

+++ code/api/src/modules/product/resolvers.js
@@ -3,16 +3,42 @@ import models from '../models'
 
 // Get product by ID
 export async function getById(parentValue, { id }) {
-    return await models.Product.findOne({ where: { id }})
+    const product = await models.Product.findOne({ where: { id }})
+
+    if(!product) {
+        // Product does not exists
+        throw new Error('The product you are looking for does not exists or has been discontinued.')
+    } else {
+        return product
+    }
 }
+
 // Get product by slug
 export async function getBySlug(parentValue, { slug }) {
-    return await models.Product.findOne({ where: { slug }})
+    const product = await models.Product.findOne({ where: { slug }})
+
+    if(!product) {
+        // Product does not exists
+        throw new Error('The product you are looking for does not exists or has been discontinued.')
+    } else {
+        return product
+    }
 }
 
 // Get all products
 export async function getAll() {
     return await models.Product.findAll()
+}
+
+// Get related products
+export async function getRelated(parentValue, { productId }) {
+    return await models.Product.findAll({
+        where: {
+            id: { [models.Sequelize.Op.not]: productId }
+        },
+        limit: 3,
+        order: [[models.Sequelize.fn('RAND')]] // mock related products by showing random products
+    })
 }
 
 // Create product
@@ -25,5 +51,12 @@ export async function create(parentValue, { name, product }) {
 
 // Delete product
 export async function remove(parentValue, { id }) {
-    return await models.Product.destroy({ where: { id }})
+    const product = await models.Product.findOne({ where: { id }})
+
+    if(!product) {
+        // Product does not exists
+        throw new Error('The product does not exists.')
+    } else {
+        return await models.Product.destroy({ where: { id }})
+    }
 }
